\name{bg.checks}
\alias{bg.checks}

\title{Some convergence diagnostics}

\description{It takes a fitted \code{bgeva} object produced by \code{bgeva()} and produces some diagnostic information about the fitting procedure.

}

\usage{

bg.checks(x)


}


\arguments{ 


\item{x}{A \code{bgeva} object produced by \code{bgeva()}.}


}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{bgeva}}
}


\keyword{binary rare events data} 
\keyword{generalized extreme value distribution} 






