\name{bgeva.gIMa}
\alias{bgeva.gIMa}

\title{Internal Function}

\description{It provides the log-likelihood, gradient and Hessian (or Fisher) information matrix for penalized or unpenalized maximum likelihood optimization.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}









