\name{bgeva.gIMb}
\alias{bgeva.gIMb}

\title{Internal Function}

\description{It provides an alternative version of the log-likelihood, gradient and Hessian (or Fisher) information matrix for penalized or unpenalized maximum likelihood optimization.}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}









