\name{print.bgeva}
\alias{print.bgeva}

\title{Print a bgeva object}

\description{The print method for a \code{bgeva} object.}

\usage{


\method{print}{bgeva}(x,...)


}


\arguments{ 


\item{x}{A \code{bgeva} object produced by \code{bgeva()}.}

\item{...}{Other arguments.}

}


\details{ 

\code{print.bgeva} prints out the family, model equation, total number of observations, chosen tail 
parameter and estimated total effective degrees of freedom for the penalized or unpenalized model. 

}



\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{bgeva}}
}


\keyword{Binary GEVA modelling} 








