\name{print.summary.bgeva}
\alias{print.summary.bgeva}

\title{Print a summary.bgeva object}

\description{The print method for a \code{summary.bgeva} object.}

\usage{


\method{print}{summary.bgeva}(x,digits = max(3, getOption("digits") - 3), 
           signif.stars = getOption("show.signif.stars"),...)


}


\arguments{ 


\item{x}{A \code{summary.bgeva} object produced by \code{summary.bgeva()}.}

\item{digits}{Number of digits printed in output.} 

\item{signif.stars}{By default significance stars are printed alongside output.} 

\item{...}{Other arguments.}

}


\details{ 

\code{print.summary.bgeva} prints model term summaries. 

}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}



\seealso{
\code{\link{summary.bgeva}}
}


\keyword{Binary GEVA modelling} 








