\name{working.comp}
\alias{working.comp}

\title{Internal Function}

\description{It efficiently calculates the working model quantities needed to implement the automatic multiple smoothing parameter 
procedure by exploiting the band structure of the weight matrix.}




\author{

Maintainer: Giampiero Marra \email{giampiero.marra@ucl.ac.uk}

}








